feat: add CLI issue state command
