+++ packages/cli/package.json
@@ -31,11 +31,12 @@
   "scripts": {
     "build": "tsc",
     "postbuild": "node scripts/ensure-shebang.mjs",
-    "test": "node ../../scripts/run-fixture-command.mjs --build @mrtdown/core --build @mrtdown/fs -- npm exec -- vitest --run"
+    "test": "node ../../scripts/run-fixture-command.mjs --build @mrtdown/core --build @mrtdown/fs --build @mrtdown/triage -- npm exec -- vitest --run"
   },
   "dependencies": {
     "@mrtdown/core": "2.0.0-alpha.24",
     "@mrtdown/fs": "2.0.0-alpha.24",
+    "@mrtdown/triage": "2.0.0-alpha.28",
     "hast-util-to-html": "^9.0.5"
   },
   "devDependencies": {
